+++ docs/source/changelog.rst
@@ -14,6 +14,8 @@ Changelog
 
 * :feature:`43` OU Model Jurek and OU Model Mudchanatongsuk added to the Stochastic Control Approach Module.
 * :support:`43` OU Model Jurek and OU Model Mudchanatongsuk documentation.
+* :feature:`44` CVine Copula and CVine Copula Strategy added to the Copula Approach Module.
+* :support:`44` CVine Copula and CVine Copula Strategy documentation.
 * :feature:`42` Options to sort pairs by zero-crossings, variance, same industry group added to the Basic Distance Strategy.
 * :support:`42` Updated Basic Distance Strategy documentation.
 * :feature:`40` Vine Copula Partner Selection Approaches added to the Copula Approach Module.